+++ changelog.md
@@ -60,4 +60,8 @@
 - Redesign and applied bootstrap (Jake)
 - Applied Allman code style (Jake)
 
+### *v.018* (Feb 11, 2017)
+- Added comments to all the history pages (Karan)
 
+### *v.019* (Feb 11, 2017)
+- Reformatted code (Karan)